+++ ewm-service/src/main/java/ru/practicum/explore/request/RequestService.java
@@ -6,15 +6,15 @@ import ru.practicum.explore.request.model.Request;
 import java.util.List;
 
 public interface RequestService {
-    List<Request> getRequestsByEvent(Integer eventId);
+    List<Request> getRequestsByEvent(Long eventId);
 
-    Request getRequestByReqId(Integer eventId, Integer reqId);
+    Request getRequestByReqId(Long eventId, Long reqId);
 
     void saveRequest(Request request);
 
-    List<ParticipationRequestDto> getRequestsByUserId(Integer userId);
+    List<ParticipationRequestDto> getRequestsByUserId(Long userId);
 
-    ParticipationRequestDto saveRequest(Integer userId, Integer eventId);
+    ParticipationRequestDto saveRequest(Long userId, Long eventId);
 
-    ParticipationRequestDto cancelRequest(Integer userId, Integer requestId);
+    ParticipationRequestDto cancelRequest(Long userId, Long requestId);
 }
